% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lalgp.R
\name{lalgp}
\alias{lalgp}
\title{Frequency graph}
\usage{
lalgp(data)
}
\arguments{
\item{data}{list data of the result of functions; \code{\link{lalgp}} or \code{\link{manual_lalgp}}.}
}
\description{
Drawing frequency graph (density plot) and analyzing the out put of
\code{\link{lalgp}} or \code{\link{manual_lalgp}}.
}
\examples{
d <- Osayama
e <- lalgp(d)
lalgp_graph(e)
}
