% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LLfreq-package.R
\docType{package}
\name{LLfreq-package}
\alias{LLfreq-package}
\alias{LLfreq}
\title{The 'LLfreq' package.}
\description{
Package to draw density curve and computation of frequency by LALGP. (Beta Version)
}
\references{
\itemize{
\item{
Tsujimori, T. (1998). Geology of the Osayama serpentinite melange in the central Chugoku Mountains,
 southwestern Japan: 320 Ma blueschist-bearing serpentinite melange beneath the Oyama ophiolite.
 Jour. Geol. Soc. Japan, 104, 213-231.
}
\item{
Riihimäki, J., and Vehtari, A. (2014). Laplace approximation for logistic Gaussian process density
estimation and regression. Bayesian analysis, 9(2), 425-448.
}
\item{
Furukawa, T. (2019). Bayesian statistical evaluation method for detrital zircon geochronology.
In Abstract of annual meeting of JpGU, Chiba, Japan, 2019.}
}
}
\author{
Tan Furukawa <rpackagetan@gmail.com>
}
